Changed from int to long long int the type of offsets relative to base pointer and stack pointer

+++ MemoryAccess.cpp
@@ -12,11 +12,11 @@ ADDRINT MemoryAccess::getAddress() const{
     return accessAddress;
 }
 
-int MemoryAccess::getSPOffset() const{
+long long int MemoryAccess::getSPOffset() const{
     return spOffset;
 }
 
-int MemoryAccess::getBPOffset() const{
+long long int MemoryAccess::getBPOffset() const{
     return bpOffset;
 }
